Miscellaneous compilation files and simplifications. (#3969)

+++ io/private/las/Summary.cpp
@@ -57,7 +57,7 @@ void Summary::addPoint(double x, double y, double z, int returnNumber)
     m_bounds.grow(x, y, z);
 
     // Returns numbers are indexed from one, but the array indexes from 0.
-    if (returnNumber >= 1 && returnNumber <= m_returnCounts.size())
+    if (returnNumber >= 1 && returnNumber <= (int)m_returnCounts.size())
         m_returnCounts[returnNumber - 1]++;
 }
